[update] refact Posts and MakePost
