+++ client/src/components/sass/landing.scss
@@ -1,47 +1,50 @@
 .landing {
-    .hero{
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-        img {
-            border-radius: 10px;
-        }
-        div {
-            margin: 0 3rem;
-            p { opacity: .7; }
-            button {
-                padding: .5rem 1rem;
-                background-color: black;
-                color: white;
-                border: none;
-                border-radius: 10px;
-            }
-        }
+  .hero {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    img {
+      border-radius: 10px;
     }
-    .blurbs, .cards {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        flex-wrap: wrap;
-        &>* {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            margin: 2rem;
-            padding: 3rem;
-            &>* {
-                margin: .5rem;
-            }
-            img {
-                height: auto;
-                width: 100px;
-                border-radius: 50%;
-            }
-        }
-        &>.card {
-            border: 2px solid lightgray;
-            border-radius: 10px;
-        }
+    div {
+      margin: 0 3rem;
+      p {
+        opacity: 0.7;
+      }
+      button {
+        padding: 0.5rem 1rem;
+        background-color: black;
+        color: white;
+        border: none;
+        border-radius: 10px;
+      }
     }
+  }
+  .blurbs,
+  .cards {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    flex-wrap: wrap;
+    & > * {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      margin: 2rem;
+      padding: 3rem;
+      & > * {
+        margin: 0.5rem;
+      }
+      img {
+        height: auto;
+        width: 100px;
+        border-radius: 50%;
+      }
+    }
+    & > .card {
+      border: 2px solid lightgray;
+      border-radius: 10px;
+    }
+  }
 }
